+++ src/global.cc
@@ -513,7 +513,7 @@ void handle_debug_options(int argc, char * argv[])
 	_log_level   = LOG_TRACE; // global in utils.h
 	try {
 	  // global in utils.h
-	  _trace_level = boost::lexical_cast<int>(argv[i + 1]);
+	  _trace_level = boost::lexical_cast<uint8_t>(argv[i + 1]);
 	}
 	catch (const boost::bad_lexical_cast& e) {
 	  throw std::logic_error(_("Argument to --trace must be an integer"));
